requirements.txt
azureml-core
numpy==1.15.4
pandas==0.23.4
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
urllib3==1.24.1